backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
PyJWT==2.9.0
python-multipart==0.0.17
aiogram==3.13.1
apscheduler==3.10.4
